word-format-tool/requirements.txt
streamlit==1.35.0
python-docx==1.1.0
pandas==2.0.3